Implemented some game classes.
Map is now GameMap because a Map is a collection type.
A GameMap holds references to all Boards and Fields.
A Board holds references to its Fields.
A Field holds a reference to its Board.
Owner is now an independent Enum that is used to refer to the owner of a
field or Board.